[trinh.dd] update product

+++ marketplace/src/main/java/com/vjshow/marketplace/service/ProductServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.vjshow.marketplace.dto.request.CompleteUploadRequest;
 import com.vjshow.marketplace.dto.request.ProductRequest;
@@ -15,15 +16,22 @@ import com.vjshow.marketplace.entity.ProductEntity;
 import com.vjshow.marketplace.enums.ProductStatusEnum;
 import com.vjshow.marketplace.enums.ProductTypeEnum;
 import com.vjshow.marketplace.exception.LogicException;
+import com.vjshow.marketplace.repository.OrderRepository;
 import com.vjshow.marketplace.repository.ProductRepository;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
 
 	private final ProductRepository productRepository;
+	
+	private final OrderRepository orderRepository;
+	
+	private final CloudFlareService r2Service;
 
 	public ProductEntity getById(Long id) {
 		return productRepository.findById(id)
@@ -57,12 +65,25 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	@Override
+	@Transactional
 	public void delete(Long id) {
 		ProductEntity product = productRepository.findById(id)
 				.orElseThrow(() -> new LogicException("NOT_FOUND", "Không tìm thấy sản phẩm"));
+		
+		 // ❌ Check nếu đã có order
+	    boolean exists = orderRepository.existsByProduct_Id(id);
+	    if (exists) {
+	        throw new LogicException("INVALID", "Sản phẩm đã có người mua, không thể xóa");
+	    }
+	    
+	    // ✅ Xóa file trên R2
+	    deleteFileIfExists(product.getPreviewUrl());
+	    deleteFileIfExists(product.getThumbnailUrl());
+	    deleteHlsByUrl(product.getHlsVideoUrl());
+		
 		productRepository.delete(product);
 	}
-
+ 
 	@Override
 	public Page<ProductEntity> getPublicProducts(String type, String keyword, int page, int size) {
 
@@ -93,5 +114,23 @@ public class ProductServiceImpl implements ProductService {
 	public List<ProductEntity> getTopProducts(ProductTypeEnum type, Long quantity) {
 		Pageable pageable = PageRequest.of(0, quantity.intValue());
 		return productRepository.findByTypeOrderByTotalSalesDesc(type, pageable);
+	}
+	
+	private void deleteFileIfExists(String url) {
+	    try {
+	        r2Service.deleteFileByUrl(url);
+	    } catch (Exception e) {
+	        // log lại nhưng KHÔNG fail transaction
+	        log.error("Delete R2 file failed: {}", url, e);
+	    }
+	}
+	
+	private void deleteHlsByUrl(String url) {
+	    try {
+	        r2Service.deleteHlsByKey(url);
+	    } catch (Exception e) {
+	        // log lại nhưng KHÔNG fail transaction
+	        log.error("Delete R2 file failed: {}", url, e);
+	    }
 	}
 }
